(db update required, see below or v5.0 release notes) Changed Portal.id type from Long to Integer, StudentAttendance.id type from Long to Integer, Announcements.id type from Long to Integer, PortalStatistics.id type from Long to Integer. alter table portal change column id id tinyint not null auto_increment; alter table runs_related_to_announcements change column announcements_fk announcements_fk mediumint not null; alter table announcements change column id id mediumint not null auto_increment; alter table portal_statistics change column id id smallint unsigned not null auto_increment; alter table student_attendance change column id id int unsigned not null auto_increment;

+++ src/main/java/org/wise/portal/service/announcement/AnnouncementService.java
@@ -47,7 +47,7 @@ public interface AnnouncementService {
 	 * 
 	 * @param <code>long</code> id
 	 */
-	void deleteAnnouncement(long id);
+	void deleteAnnouncement(Integer id);
 	
 	/**
 	 * updates an Announcement in the data store
@@ -57,7 +57,7 @@ public interface AnnouncementService {
 	 * @return <code>Announcement</code>
 	 * @throws <code>ObjectNotFoundException</code>
 	 */
-	Announcement updateAnnouncement(long id, AnnouncementParameters params) throws ObjectNotFoundException;
+	Announcement updateAnnouncement(Integer id, AnnouncementParameters params) throws ObjectNotFoundException;
 	
 	/**
 	 * retrieves the Announcement with the given Id from the data store
@@ -66,5 +66,5 @@ public interface AnnouncementService {
 	 * @return <code>Announcement</code>
 	 * @throws <code>ObjectNotFoundException</code>
 	 */
-	Announcement retrieveById(long id) throws ObjectNotFoundException;
+	Announcement retrieveById(Integer id) throws ObjectNotFoundException;
 }
